+++ backend/main.py
@@ -27,7 +27,7 @@ url_prefix = "/api/v1"
 
 app = FastAPI(
     title="StickerMap API",
-    version="1.11.0",
+    version="1.12.0",
     debug=True,
     docs_url=url_prefix + "/docs",
     redoc_url=url_prefix + "/redoc",